test/requirements.txt
Flask==2.3.4
numpy
tensorflow==2.13.0
Pillow
